Segundo commit
Se añadio el calculo del tiempo en conseguir la palabra, falta revisar los falsos positivos (casos donde se encuentra parte de la palabra) y antes de cerrar la funcion de busqueda hay que regresar el estado de las letras (enUso) a falso y hay que revisar que con mas casos porque en ese de los falsos positivos hay que regresar el estado de las las letras que se tenían en uso a falso.

+++ Palabra.java
@@ -12,7 +12,7 @@ public class Palabra {
     String info;
     Palabra sig;
     boolean encontrada;
-    float tiempo;
+    long tiempo;
 
     public Palabra(String info) {
         this.info = info;
